+++ src/test/java/fi/metatavu/keycloak/scim/server/test/ScimClient.java
@@ -1,4 +1,4 @@
-package fi.metatavu.keycloak.scim.server;
+package fi.metatavu.keycloak.scim.server.test;
 
 import fi.metatavu.keycloak.scim.server.test.client.ApiClient;
 import fi.metatavu.keycloak.scim.server.test.client.ApiException;
