+++ src/test/java/fi/metatavu/keycloak/scim/server/test/tests/functional/OrganizationUserPatchTestsIT.java
@@ -1,10 +1,14 @@
-package fi.metatavu.keycloak.scim.server;
+package fi.metatavu.keycloak.scim.server.test.tests.functional;
 
 import dasniko.testcontainers.keycloak.KeycloakContainer;
+import fi.metatavu.keycloak.scim.server.test.tests.AbstractOrganizationScimTest;
+import fi.metatavu.keycloak.scim.server.test.ScimClient;
+import fi.metatavu.keycloak.scim.server.test.TestConsts;
 import fi.metatavu.keycloak.scim.server.test.client.ApiException;
 import fi.metatavu.keycloak.scim.server.test.client.model.PatchRequest;
 import fi.metatavu.keycloak.scim.server.test.client.model.PatchRequestOperationsInner;
 import fi.metatavu.keycloak.scim.server.test.client.model.User;
+import fi.metatavu.keycloak.scim.server.test.utils.KeycloakTestUtils;
 import org.junit.jupiter.api.Test;
 import org.keycloak.representations.idm.UserRepresentation;
 import org.testcontainers.junit.jupiter.Container;
@@ -18,16 +22,15 @@ import static org.junit.jupiter.api.Assertions.*;
  * Tests for SCIM 2.0 User create endpoint
  */
 @Testcontainers
-public class RealmUserPatchTestsIT extends AbstractRealmScimTest {
+public class OrganizationUserPatchTestsIT extends AbstractOrganizationScimTest {
 
     @Container
     protected static final KeycloakContainer keycloakContainer = new KeycloakContainer("quay.io/keycloak/keycloak:26.1.2")
-        .withNetwork(network)
-        .withNetworkAliases("scim-keycloak")
-        .withEnv("SCIM_AUTHENTICATION_MODE", "KEYCLOAK")
-        .withProviderLibsFrom(KeycloakTestUtils.getBuildProviders())
-        .withRealmImportFile("kc-test.json")
-        .withLogConsumer(outputFrame -> System.out.printf("KEYCLOAK: %s", outputFrame.getUtf8String()));
+            .withNetwork(network)
+            .withNetworkAliases("scim-keycloak")
+            .withProviderLibsFrom(KeycloakTestUtils.getBuildProviders())
+            .withRealmImportFiles("kc-organizations.json", "kc-external.json")
+            .withLogConsumer(outputFrame -> System.out.printf("KEYCLOAK: %s", outputFrame.getUtf8String()));
 
     @Override
     protected KeycloakContainer getKeycloakContainer() {
@@ -36,7 +39,7 @@ public class RealmUserPatchTestsIT extends AbstractRealmScimTest {
 
     @Test
     void testActivateAndDeactivateUser() throws ApiException {
-        ScimClient scimClient = getAuthenticatedScimClient();
+        ScimClient scimClient = getAuthenticatedScimClient(TestConsts.ORGANIZATION_1_ID);
 
         // Create an active user
         User user = new User();
@@ -49,7 +52,7 @@ public class RealmUserPatchTestsIT extends AbstractRealmScimTest {
         assertNotNull(created.getActive());
         assertTrue(created.getActive());
 
-        UserRepresentation createdRealmUser = findRealmUser(TestConsts.TEST_REALM, created.getId());
+        UserRepresentation createdRealmUser = findRealmUser(TestConsts.ORGANIZATIONS_REALM, created.getId());
         assertNotNull(createdRealmUser);
         assertTrue(createdRealmUser.isEnabled());
 
@@ -67,7 +70,7 @@ public class RealmUserPatchTestsIT extends AbstractRealmScimTest {
         assertNotNull(deactivated.getActive());
         assertFalse(deactivated.getActive());
 
-        UserRepresentation deactivatedRealmUser = findRealmUser(TestConsts.TEST_REALM, created.getId());
+        UserRepresentation deactivatedRealmUser = findRealmUser(TestConsts.ORGANIZATIONS_REALM, created.getId());
         assertNotNull(deactivatedRealmUser);
         assertFalse(deactivatedRealmUser.isEnabled());
 
@@ -84,17 +87,17 @@ public class RealmUserPatchTestsIT extends AbstractRealmScimTest {
         assertNotNull(activated.getActive());
         assertTrue(activated.getActive());
 
-        UserRepresentation activatedRealmUser = findRealmUser(TestConsts.TEST_REALM, created.getId());
+        UserRepresentation activatedRealmUser = findRealmUser(TestConsts.ORGANIZATIONS_REALM, created.getId());
         assertNotNull(activatedRealmUser);
         assertTrue(activatedRealmUser.isEnabled());
 
         // Cleanup
-        deleteRealmUser(TestConsts.TEST_REALM, created.getId());
+        deleteRealmUser(TestConsts.ORGANIZATIONS_REALM, created.getId());
     }
 
     @Test
     void testPatchAttributes() throws ApiException {
-        ScimClient scimClient = getAuthenticatedScimClient();
+        ScimClient scimClient = getAuthenticatedScimClient(TestConsts.ORGANIZATION_1_ID);
 
         // Create user
         User user = new User();
@@ -156,7 +159,7 @@ public class RealmUserPatchTestsIT extends AbstractRealmScimTest {
         assertEquals("en_US", patchedAgain.getAdditionalProperty("preferredLanguage"));
 
         // Cleanup
-        deleteRealmUser(TestConsts.TEST_REALM, created.getId());
+        deleteRealmUser(TestConsts.ORGANIZATIONS_REALM, created.getId());
     }
 
 }
