+++ src/test/java/fi/metatavu/keycloak/scim/server/test/tests/functional/OrganizationUserUpdateTestsIT.java
@@ -1,8 +1,12 @@
-package fi.metatavu.keycloak.scim.server;
+package fi.metatavu.keycloak.scim.server.test.tests.functional;
 
 import dasniko.testcontainers.keycloak.KeycloakContainer;
+import fi.metatavu.keycloak.scim.server.test.tests.AbstractOrganizationScimTest;
+import fi.metatavu.keycloak.scim.server.test.ScimClient;
+import fi.metatavu.keycloak.scim.server.test.TestConsts;
 import fi.metatavu.keycloak.scim.server.test.client.ApiException;
 import fi.metatavu.keycloak.scim.server.test.client.model.User;
+import fi.metatavu.keycloak.scim.server.test.utils.KeycloakTestUtils;
 import org.junit.jupiter.api.Test;
 import org.keycloak.representations.idm.UserRepresentation;
 import org.testcontainers.junit.jupiter.Container;
@@ -16,16 +20,15 @@ import static org.junit.jupiter.api.Assertions.*;
  * Tests for SCIM 2.0 User update (PUT) endpoint
  */
 @Testcontainers
-public class RealmUserUpdateTestsIT extends AbstractRealmScimTest {
+public class OrganizationUserUpdateTestsIT extends AbstractOrganizationScimTest {
 
     @Container
     protected static final KeycloakContainer keycloakContainer = new KeycloakContainer("quay.io/keycloak/keycloak:26.1.2")
-        .withNetwork(network)
-        .withNetworkAliases("scim-keycloak")
-        .withEnv("SCIM_AUTHENTICATION_MODE", "KEYCLOAK")
-        .withProviderLibsFrom(KeycloakTestUtils.getBuildProviders())
-        .withRealmImportFile("kc-test.json")
-        .withLogConsumer(outputFrame -> System.out.printf("KEYCLOAK: %s", outputFrame.getUtf8String()));
+            .withNetwork(network)
+            .withNetworkAliases("scim-keycloak")
+            .withProviderLibsFrom(KeycloakTestUtils.getBuildProviders())
+            .withRealmImportFiles("kc-organizations.json", "kc-external.json")
+            .withLogConsumer(outputFrame -> System.out.printf("KEYCLOAK: %s", outputFrame.getUtf8String()));
 
     @Override
     protected KeycloakContainer getKeycloakContainer() {
@@ -34,7 +37,7 @@ public class RealmUserUpdateTestsIT extends AbstractRealmScimTest {
 
     @Test
     void testReplaceUser() throws ApiException {
-        ScimClient scimClient = getAuthenticatedScimClient();
+        ScimClient scimClient = getAuthenticatedScimClient(TestConsts.ORGANIZATION_1_ID);
 
         // Create initial user
         User user = new User();
@@ -79,7 +82,7 @@ public class RealmUserUpdateTestsIT extends AbstractRealmScimTest {
         assertFalse(updated.getActive());
 
         // Also verify state in Keycloak
-        UserRepresentation realmUser = findRealmUser(TestConsts.TEST_REALM, userId);
+        UserRepresentation realmUser = findRealmUser(TestConsts.ORGANIZATIONS_REALM, userId);
         assertNotNull(realmUser);
         assertEquals("replace-user", realmUser.getUsername());
         assertEquals("Replaced", realmUser.getFirstName());
@@ -91,12 +94,12 @@ public class RealmUserUpdateTestsIT extends AbstractRealmScimTest {
         assertFalse(realmUser.isEnabled());
 
         // Clean up
-        deleteRealmUser(TestConsts.TEST_REALM, userId);
+        deleteRealmUser(TestConsts.ORGANIZATIONS_REALM, userId);
     }
 
     @Test
     void testReplaceNonExistentUserReturnsNotFound() {
-        ScimClient scimClient = getAuthenticatedScimClient();
+        ScimClient scimClient = getAuthenticatedScimClient(TestConsts.ORGANIZATION_1_ID);
 
         User replacement = new User();
         replacement.setUserName("ghost");
@@ -112,7 +115,7 @@ public class RealmUserUpdateTestsIT extends AbstractRealmScimTest {
 
     @Test
     void testReplaceUserWithConflictingUserNameReturnsConflict() throws ApiException {
-        ScimClient scimClient = getAuthenticatedScimClient();
+        ScimClient scimClient = getAuthenticatedScimClient(TestConsts.ORGANIZATION_1_ID);
 
         // Create two users
         User userA = new User();
@@ -137,8 +140,8 @@ public class RealmUserUpdateTestsIT extends AbstractRealmScimTest {
         assertEquals(409, exception.getCode());
 
         // Clean up
-        deleteRealmUser(TestConsts.TEST_REALM, createdA.getId());
-        deleteRealmUser(TestConsts.TEST_REALM, createdB.getId());
+        deleteRealmUser(TestConsts.ORGANIZATIONS_REALM, createdA.getId());
+        deleteRealmUser(TestConsts.ORGANIZATIONS_REALM, createdB.getId());
     }
 
 }
