+++ src/test/java/fi/metatavu/keycloak/scim/server/test/tests/functional/RealmResourceTypesTestsIT.java
@@ -1,8 +1,12 @@
-package fi.metatavu.keycloak.scim.server;
+package fi.metatavu.keycloak.scim.server.test.tests.functional;
 
 import dasniko.testcontainers.keycloak.KeycloakContainer;
+import fi.metatavu.keycloak.scim.server.test.tests.AbstractRealmScimTest;
+import fi.metatavu.keycloak.scim.server.test.ScimClient;
 import fi.metatavu.keycloak.scim.server.test.client.ApiException;
-import fi.metatavu.keycloak.scim.server.test.client.model.*;
+import fi.metatavu.keycloak.scim.server.test.client.model.ResourceType;
+import fi.metatavu.keycloak.scim.server.test.client.model.ResourceTypeListResponse;
+import fi.metatavu.keycloak.scim.server.test.utils.KeycloakTestUtils;
 import org.junit.jupiter.api.Test;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
@@ -12,28 +16,28 @@ import java.util.List;
 import static org.junit.jupiter.api.Assertions.*;
 
 /**
- * Tests for SCIM 2.0 Schemas endpoint
+ * Tests for realm server SCIM 2.0 user find (GET /Users/{id}) endpoint
  */
 @Testcontainers
-public class OrganizationResourceTypesTestsIT extends AbstractOrganizationScimTest {
+public class RealmResourceTypesTestsIT extends AbstractRealmScimTest {
 
     @Container
     protected static final KeycloakContainer keycloakContainer = new KeycloakContainer("quay.io/keycloak/keycloak:26.1.2")
-            .withNetwork(network)
-            .withNetworkAliases("scim-keycloak")
-            .withProviderLibsFrom(KeycloakTestUtils.getBuildProviders())
-            .withRealmImportFiles("kc-organizations.json", "kc-external.json")
-            .withLogConsumer(outputFrame -> System.out.printf("KEYCLOAK: %s", outputFrame.getUtf8String()));
+        .withNetwork(network)
+        .withNetworkAliases("scim-keycloak")
+        .withEnv("SCIM_AUTHENTICATION_MODE", "KEYCLOAK")
+        .withProviderLibsFrom(KeycloakTestUtils.getBuildProviders())
+        .withRealmImportFile("kc-test.json")
+        .withLogConsumer(outputFrame -> System.out.printf("KEYCLOAK: %s", outputFrame.getUtf8String()));
 
     @Override
     protected KeycloakContainer getKeycloakContainer() {
         return keycloakContainer;
     }
 
-
     @Test
-    void testResourceTypesOrg1() throws ApiException {
-        ScimClient scimClient = getAuthenticatedScimClient(TestConsts.ORGANIZATION_1_ID);
+    void testResourceTypes() throws ApiException {
+        ScimClient scimClient = getAuthenticatedScimClient();
 
         ResourceTypeListResponse listResponse = scimClient.getResourceTypes();
         List<ResourceType> resourceTypes = listResponse.getResources();
@@ -44,8 +48,8 @@ public class OrganizationResourceTypesTestsIT extends AbstractOrganizationScimTe
     }
 
     @Test
-    void testUserResourceTypeOrg1() throws ApiException {
-        ScimClient scimClient = getAuthenticatedScimClient(TestConsts.ORGANIZATION_1_ID);
+    void testUserResourceType() throws ApiException {
+        ScimClient scimClient = getAuthenticatedScimClient();
 
         ResourceType resourceType = scimClient.findResourceType("User");
         assertNotNull(resourceType);
@@ -53,8 +57,8 @@ public class OrganizationResourceTypesTestsIT extends AbstractOrganizationScimTe
     }
 
     @Test
-    void testGroupResourceTypeOrg1() throws ApiException {
-        ScimClient scimClient = getAuthenticatedScimClient(TestConsts.ORGANIZATION_1_ID);
+    void testGroupResourceType() throws ApiException {
+        ScimClient scimClient = getAuthenticatedScimClient();
 
         ResourceType resourceType = scimClient.findResourceType("Group");
         assertNotNull(resourceType);
@@ -70,10 +74,10 @@ public class OrganizationResourceTypesTestsIT extends AbstractOrganizationScimTe
      * @param description description
      */
     private void assertResourceType(
-            ResourceType resourceType,
-            String id,
-            String path,
-            String description
+        ResourceType resourceType,
+        String id,
+        String path,
+        String description
     ) {
         assertArrayEquals(new String[] { "urn:ietf:params:scim:schemas:core:2.0:ResourceType" }, resourceType.getSchemas().toArray());
         assertEquals(id, resourceType.getId());
@@ -85,7 +89,7 @@ public class OrganizationResourceTypesTestsIT extends AbstractOrganizationScimTe
         assertEquals(0, resourceType.getSchemaExtensions().size());
         assertNotNull(resourceType.getMeta());
         assertEquals(id, resourceType.getMeta().getResourceType());
-        assertEquals(getScimUri(TestConsts.ORGANIZATION_1_ID).resolve(String.format("ResourceTypes/%s", id)), resourceType.getMeta().getLocation());
+        assertEquals(getScimUri().resolve(String.format("ResourceTypes/%s", id)), resourceType.getMeta().getLocation());
     }
 
 }
