+++ src/test/java/fi/metatavu/keycloak/scim/server/test/tests/functional/RealmUserFindTestsIT.java
@@ -1,9 +1,12 @@
-package fi.metatavu.keycloak.scim.server;
+package fi.metatavu.keycloak.scim.server.test.tests.functional;
 
 import dasniko.testcontainers.keycloak.KeycloakContainer;
+import fi.metatavu.keycloak.scim.server.test.tests.AbstractRealmScimTest;
+import fi.metatavu.keycloak.scim.server.test.ScimClient;
+import fi.metatavu.keycloak.scim.server.test.TestConsts;
 import fi.metatavu.keycloak.scim.server.test.client.ApiException;
 import fi.metatavu.keycloak.scim.server.test.client.model.User;
-import jakarta.ws.rs.NotFoundException;
+import fi.metatavu.keycloak.scim.server.test.utils.KeycloakTestUtils;
 import org.junit.jupiter.api.Test;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
@@ -13,10 +16,10 @@ import java.util.List;
 import static org.junit.jupiter.api.Assertions.*;
 
 /**
- * Tests for SCIM 2.0 User delete endpoint
+ * Tests for SCIM 2.0 user find (GET /Users/{id}) endpoint
  */
 @Testcontainers
-public class RealmUserDeleteTestsIT extends AbstractRealmScimTest {
+public class RealmUserFindTestsIT extends AbstractRealmScimTest {
 
     @Container
     protected static final KeycloakContainer keycloakContainer = new KeycloakContainer("quay.io/keycloak/keycloak:26.1.2")
@@ -33,43 +36,44 @@ public class RealmUserDeleteTestsIT extends AbstractRealmScimTest {
     }
 
     @Test
-    void testDeleteUser() throws ApiException {
+    void testFindUserById() throws ApiException {
         ScimClient scimClient = getAuthenticatedScimClient();
 
         // Create user
         User user = new User();
-        user.setUserName("delete-me");
+        user.setUserName("find-me");
         user.setActive(true);
         user.setSchemas(List.of("urn:ietf:params:scim:schemas:core:2.0:User"));
-        user.setName(getName("Delete", "Me"));
-        user.setEmails(getEmails("delete.me@example.com"));
+        user.setName(getName("Find", "Me"));
+        user.setEmails(getEmails("find.me@example.com"));
 
         User created = scimClient.createUser(user);
         assertNotNull(created);
+        String userId = created.getId();
 
-        // Delete user
-        scimClient.deleteUser(created.getId());
+        // Find the user
+        User found = scimClient.findUser(userId);
+        assertNotNull(found);
+        assertEquals(userId, found.getId());
+        assertEquals("find-me", found.getUserName());
+        assertNotNull(found.getName());
+        assertEquals("Find", found.getName().getGivenName());
+        assertEquals("Me", found.getName().getFamilyName());
+        assertNotNull(found.getEmails());
+        assertEquals("find.me@example.com", found.getEmails().getFirst().getValue());
 
-        // Try to fetch the user to confirm deletion
-        ApiException exception = assertThrows(ApiException.class, () ->
-            scimClient.findUser(created.getId())
-        );
-
-        assertEquals(404, exception.getCode());
-
-        NotFoundException notFoundException = assertThrows(NotFoundException.class, () ->
-            findRealmUser(TestConsts.TEST_REALM, created.getId())
-        );
+        // Clean up
+        deleteRealmUser(TestConsts.TEST_REALM, userId);
     }
 
     @Test
-    void testDeleteNonexistentUserReturns404() {
+    void testFindUserNotFound() {
         ScimClient scimClient = getAuthenticatedScimClient();
 
-        String nonexistentId = "nonexistent-user-id";
+        String fakeId = "non-existent-id";
 
         ApiException exception = assertThrows(ApiException.class, () ->
-                scimClient.deleteUser(nonexistentId)
+                scimClient.findUser(fakeId)
         );
 
         assertEquals(404, exception.getCode());
